Move School subform into own component
