+++ src/components/RegisterForm/RegisterForm.tsx
@@ -1,22 +1,20 @@
-import {useMutation, useQuery} from '@tanstack/react-query'
+import {useMutation} from '@tanstack/react-query'
 import axios from 'axios'
 import {useRouter} from 'next/router'
-import {FC, useEffect, useRef} from 'react'
+import {FC} from 'react'
 import {SubmitHandler, useForm} from 'react-hook-form'
 
 import styles from '@/components/FormItems/Form.module.scss'
-import {FormAutocomplete} from '@/components/FormItems/FormAutocomplete/FormAutocomplete'
 import {FormCheckbox} from '@/components/FormItems/FormCheckbox/FormCheckbox'
 import {FormInput} from '@/components/FormItems/FormInput/FormInput'
-import {FormSelect, SelectOption} from '@/components/FormItems/FormSelect/FormSelect'
-import {IGrade} from '@/types/api/competition'
+import {SelectOption} from '@/components/FormItems/FormSelect/FormSelect'
 import {IGeneralPostResponse} from '@/types/api/general'
-import {ISchool} from '@/types/api/personal'
 import {useSeminarInfo} from '@/utils/useSeminarInfo'
 
 import {Button} from '../Clickable/Clickable'
+import {SchoolSubForm} from '../SchoolSubForm/SchoolSubForm'
 
-type RegisterFormValues = {
+export type RegisterFormValues = {
   email?: string
   password1?: string
   password2?: string
@@ -50,7 +48,6 @@ const defaultValues: RegisterFormValues = {
 
 export const RegisterForm: FC = () => {
   const {handleSubmit, control, watch, setValue, getValues} = useForm<RegisterFormValues>({defaultValues})
-  const [school_not_found, without_school] = watch(['school_not_found', 'without_school'])
 
   const scrollToTop = () => {
     window.scrollTo({
@@ -59,57 +56,7 @@ export const RegisterForm: FC = () => {
     })
   }
 
-  const otherSchoolItem = useRef<SelectOption>()
-  const withoutSchoolItem = useRef<SelectOption>()
-
   const router = useRouter()
-
-  // načítanie ročníkov z BE, ktorými vyplníme FormSelect s ročníkmi
-  const {data: gradesData} = useQuery({
-    queryKey: ['competition', 'grade'],
-    queryFn: () => axios.get<IGrade[]>(`/api/competition/grade`),
-  })
-  const grades = gradesData?.data ?? []
-  const gradeItems: SelectOption[] = grades.map(({id, name}) => ({id, label: name}))
-
-  // načítanie škôl z BE, ktorými vyplníme FormAutocomplete so školami
-  const {data: schoolsData} = useQuery({
-    queryKey: ['personal', 'schools'],
-    queryFn: () => axios.get<ISchool[]>(`/api/personal/schools`),
-  })
-  const schools = schoolsData?.data ?? []
-  const allSchoolItems: SelectOption[] = schools.map(({code, city, name, street}) => ({
-    id: code,
-    label: city ? `${name} ${street}, ${city}` : name,
-  }))
-  const emptySchoolItems = allSchoolItems.filter(({id}) => [0, 1].includes(id))
-  otherSchoolItem.current = emptySchoolItems.find(({id}) => id === 0)
-  withoutSchoolItem.current = emptySchoolItems.find(({id}) => id === 1)
-  const schoolItems = allSchoolItems.filter(({id}) => ![0, 1].includes(id))
-
-  // predvyplnenie/zmazania hodnôt pri zakliknutí checkboxu pre užívateľa po škole
-  useEffect(() => {
-    if (without_school) {
-      setValue('school', withoutSchoolItem.current)
-      setValue('grade', 13)
-      setValue('school_not_found', false)
-    } else {
-      setValue('school', null)
-      setValue('grade', '')
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [without_school])
-
-  // predvyplnenie/zmazania hodnôt pri zakliknutí checkboxu pre neznámu školu
-  useEffect(() => {
-    if (school_not_found) {
-      setValue('school', otherSchoolItem.current)
-    } else if (!without_school) {
-      setValue('school', null)
-      setValue('grade', '')
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [school_not_found])
 
   const {seminar} = useSeminarInfo()
 
@@ -189,37 +136,7 @@ export const RegisterForm: FC = () => {
         />
         <FormInput control={control} name="first_name" label="krstné meno*" rules={requiredRule} />
         <FormInput control={control} name="last_name" label="priezvisko*" rules={requiredRule} />
-        <FormCheckbox control={control} name="without_school" label="nie som študent základnej ani strednej školy." />
-        <FormAutocomplete
-          control={control}
-          name="school"
-          label="škola*"
-          options={school_not_found || without_school ? emptySchoolItems : schoolItems}
-          disabled={!schoolItems.length || school_not_found || without_school}
-          rules={requiredRule}
-        />
-        <FormCheckbox
-          control={control}
-          name="school_not_found"
-          label="moja škola sa v zozname nenachádza."
-          disabled={without_school}
-        />
-        {school_not_found && (
-          <FormInput
-            control={control}
-            name="new_school_description"
-            label="povedz nám, na akú školu chodíš, aby sme ti ju mohli pridať*"
-            rules={school_not_found ? requiredRule : {}}
-          />
-        )}
-        <FormSelect
-          control={control}
-          name="grade"
-          label="ročník*"
-          options={gradeItems.filter(({id}) => id !== 13 || without_school)}
-          disabled={!gradeItems.length || without_school}
-          rules={requiredRule}
-        />
+        <SchoolSubForm control={control} watch={watch} setValue={setValue} />
         <FormInput control={control} name="phone" label="telefónne číslo" rules={phoneRule} />
         <FormInput control={control} name="parent_phone" label="telefónne číslo na rodiča" rules={phoneRule} />
         <FormCheckbox
