+++ src/components/SchoolSubForm/SchoolSubForm.tsx
@@ -1,0 +1,112 @@
+import {useQuery} from '@tanstack/react-query'
+import axios from 'axios'
+import {FC, useEffect, useRef} from 'react'
+import {Control, UseFormSetValue, UseFormWatch} from 'react-hook-form'
+
+import styles from '@/components/FormItems/Form.module.scss'
+import {IGrade} from '@/types/api/competition'
+import {ISchool} from '@/types/api/personal'
+
+import {FormAutocomplete} from '../FormItems/FormAutocomplete/FormAutocomplete'
+import {FormCheckbox} from '../FormItems/FormCheckbox/FormCheckbox'
+import {FormInput} from '../FormItems/FormInput/FormInput'
+import {FormSelect, SelectOption} from '../FormItems/FormSelect/FormSelect'
+import {RegisterFormValues} from '../RegisterForm/RegisterForm'
+
+type SchoolSubFormProps = {
+  control: Control<RegisterFormValues, unknown>
+  watch: UseFormWatch<RegisterFormValues>
+  setValue: UseFormSetValue<RegisterFormValues>
+}
+
+export const SchoolSubForm: FC<SchoolSubFormProps> = ({control, watch, setValue}) => {
+  const [school_not_found, without_school] = watch(['school_not_found', 'without_school'])
+
+  const otherSchoolItem = useRef<SelectOption>()
+  const withoutSchoolItem = useRef<SelectOption>()
+
+  // načítanie ročníkov z BE, ktorými vyplníme FormSelect s ročníkmi
+  const {data: gradesData} = useQuery({
+    queryKey: ['competition', 'grade'],
+    queryFn: () => axios.get<IGrade[]>(`/api/competition/grade`),
+  })
+  const grades = gradesData?.data ?? []
+  const gradeItems: SelectOption[] = grades.map(({id, name}) => ({id, label: name}))
+
+  // načítanie škôl z BE, ktorými vyplníme FormAutocomplete so školami
+  const {data: schoolsData} = useQuery({
+    queryKey: ['personal', 'schools'],
+    queryFn: () => axios.get<ISchool[]>(`/api/personal/schools`),
+  })
+  const schools = schoolsData?.data ?? []
+  const allSchoolItems: SelectOption[] = schools.map(({code, city, name, street}) => ({
+    id: code,
+    label: city ? `${name} ${street}, ${city}` : name,
+  }))
+  const emptySchoolItems = allSchoolItems.filter(({id}) => [0, 1].includes(id))
+  otherSchoolItem.current = emptySchoolItems.find(({id}) => id === 0)
+  withoutSchoolItem.current = emptySchoolItems.find(({id}) => id === 1)
+  const schoolItems = allSchoolItems.filter(({id}) => ![0, 1].includes(id))
+
+  // predvyplnenie/zmazania hodnôt pri zakliknutí checkboxu pre užívateľa po škole
+  useEffect(() => {
+    if (without_school) {
+      setValue('school', withoutSchoolItem.current)
+      setValue('grade', 13)
+      setValue('school_not_found', false)
+    } else {
+      setValue('school', null)
+      setValue('grade', '')
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [without_school])
+
+  // predvyplnenie/zmazania hodnôt pri zakliknutí checkboxu pre neznámu školu
+  useEffect(() => {
+    if (school_not_found) {
+      setValue('school', otherSchoolItem.current)
+    } else if (!without_school) {
+      setValue('school', null)
+      setValue('grade', '')
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [school_not_found])
+
+  const requiredRule = {required: '* Toto pole nemôže byť prázdne.'}
+  return (
+    <div className={styles.form}>
+      <FormCheckbox control={control} name="without_school" label="nie som študent základnej ani strednej školy." />
+      <FormAutocomplete
+        control={control}
+        name="school"
+        label="škola*"
+        options={school_not_found || without_school ? emptySchoolItems : schoolItems}
+        disabled={!schoolItems.length || school_not_found || without_school}
+        rules={requiredRule}
+      />
+      <FormCheckbox
+        control={control}
+        name="school_not_found"
+        label="moja škola sa v zozname nenachádza."
+        disabled={without_school}
+      />
+      {school_not_found && (
+        <FormInput
+          control={control}
+          name="new_school_description"
+          label="povedz nám, na akú školu chodíš, aby sme ti ju mohli pridať*"
+          rules={school_not_found ? requiredRule : {}}
+          sx={{mb: '1rem'}}
+        />
+      )}
+      <FormSelect
+        control={control}
+        name="grade"
+        label="ročník*"
+        options={gradeItems.filter(({id}) => id !== 13 || without_school)}
+        disabled={!gradeItems.length || without_school}
+        rules={requiredRule}
+      />
+    </div>
+  )
+}
